Port agent to python3
Added compatibility check with few exception not resolved yet.

Following compatibility issues resolved
 * Print statements
 * rount() method
 * Exception handling
 * __metaclass__ handling
 * range() method
 * Stderr output

Following are still to be fixed.

 * Absolute import in all files
 * Old division
 * long conversion

+++ python/src/pysdk/pysdk/vmconfig.py
@@ -151,7 +151,7 @@ def CreateDefaultSpec(name="Dummy VM", memory=128, guest="otherGuest",
                       datastoreName=None, policy=None):
     cspec = vim.vm.ConfigSpec()
     cspec.annotation = annotation
-    cspec.memoryMB = long(memory)
+    cspec.memoryMB = memory
     cspec.guestId = guest
     cspec.name = name
     cspec.numCPUs = cpus
